Fix start and stop scripts for Play 2.2.0
The dist task of Play 2.2.0 no longer creates target/start, instead it creates a script in target/universal/stage/bin. Also it puts the RUNNING_PID file in target/universal/stage when the generated script is used to start the app.

This patch fixes the script generation to work with Play 2.2.0, but does not address backward compatibility.